+++ pkg/querier/queryrange/roundtrip_test.go
@@ -103,14 +103,20 @@ func TestTripperware_Metrics(t *testing.T) {
 	tests := map[string]struct {
 		path                    string
 		expectedNotAlignedCount int
+		stepAlignEnabled        bool
 	}{
 		"start/end is aligned to step": {
 			path:                    "/api/v1/query_range?query=up&start=1536673680&end=1536716880&step=120",
 			expectedNotAlignedCount: 0,
 		},
-		"start/end is not aligned to step": {
+		"start/end is not aligned to step, aligning disabled": {
 			path:                    "/api/v1/query_range?query=up&start=1536673680&end=1536716880&step=7",
 			expectedNotAlignedCount: 1,
+		},
+		"start/end is not aligned to step, aligning enabled": {
+			path:                    "/api/v1/query_range?query=up&start=1536673680&end=1536716880&step=7",
+			expectedNotAlignedCount: 1,
+			stepAlignEnabled:        true,
 		},
 	}
 
@@ -135,7 +141,7 @@ func TestTripperware_Metrics(t *testing.T) {
 	for testName, testData := range tests {
 		t.Run(testName, func(t *testing.T) {
 			reg := prometheus.NewPedanticRegistry()
-			tw, _, err := NewTripperware(Config{},
+			tw, _, err := NewTripperware(Config{AlignQueriesWithStep: testData.stepAlignEnabled},
 				log.NewNopLogger(),
 				mockLimits{},
 				PrometheusCodec,